% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_settings.R
\name{ez_gg}
\alias{ez_gg}
\title{Annie's Quick ggplot Settings}
\usage{
ez_gg(base_size = 12, base_family = "Times New Roman")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}
}
\value{
A ggplot2 object
}
\description{
Applies formatting for  publication-ready figs
}
